requirements.txt
psycopg[binary]==3.2.3
python-dotenv==1.0.1
tqdm==4.66.4
pydantic==2.6.4
requests==2.32.3
trafilatura==1.8.0
beautifulsoup4==4.12.3
pypdf==5.0.1
pdfplumber==0.11.7
pytesseract==0.3.13
pdf2image==1.17.0
ebooklib==0.18
git+https://github.com/m-bain/whisperx.git
ffmpeg-python==0.2.0
numpy>=2.0.2,<2.1.0
sentence-transformers==3.0.1
rapidfuzz==3.9.7
playwright==1.48.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.39.0
